refactor: move data handlers to utils.
